🧭 Remove classic submissions route
