+++ ee/sites/src/components/SubmissionList.tsx
@@ -1,6 +1,6 @@
-import type { SubmissionListingSiteContext } from '../routes/$siteName.submissions-classic/site-context.format.server.js';
+import type { SubmissionListingSiteContext } from '../routes/$siteName.submissions._index/site-context.format.server.js';
 import { SubmissionListItem } from './SubmissionListItem.js';
-import type { AugmentedSubmissionListingItem } from '../routes/$siteName.submissions-classic/types.js';
+import type { AugmentedSubmissionListingItem } from '../routes/$siteName.submissions._index/types.js';
 
 export function SubmissionList({
   site,
