--- ee/sites/src/routes/$siteName.submissions-classic/CollectionSelect.tsx
@@ -1,59 +1,0 @@
-import type { FetcherWithComponents } from 'react-router';
-import { useNavigate } from 'react-router';
-import type { JobDTO } from '@curvenote/common';
-import { primitives } from '@curvenote/scms-core';
-import type { SubmissionListingPage } from './types.js';
-import type { CollectionFilterOption } from './collections.format.server.js';
-
-export function CollectionSelect({
-  collections,
-  fetcher,
-  defaultValue,
-}: {
-  fetcher: FetcherWithComponents<{
-    submissions?: SubmissionListingPage;
-    error?: string;
-    jobs: { items: JobDTO[] };
-  }>;
-  collections: CollectionFilterOption[];
-  defaultValue?: string;
-}) {
-  const navigate = useNavigate();
-
-  const handleCollectionSelect = async (collection: string) => {
-    const formData = new FormData();
-    formData.append('intent', 'filter-collection');
-    formData.append('collection', collection);
-    fetcher.submit(formData, { method: 'post' });
-
-    // Update URL without page reload
-    const searchParams = new URLSearchParams(window.location.search);
-    if (collection === 'all') {
-      searchParams.delete('collection');
-    } else {
-      searchParams.set('collection', collection);
-    }
-    console.log('navigate', searchParams.toString());
-    navigate(`?${searchParams.toString()}`);
-  };
-
-  return (
-    <div className="flex items-center gap-4 mb-2">
-      <div>Show:</div>
-      <primitives.Select
-        placeholder="All Collections"
-        ariaLabel="select collections"
-        defaultValue={defaultValue ?? 'all'}
-        onValueChange={handleCollectionSelect}
-      >
-        <primitives.SelectItem value="all">All Collections</primitives.SelectItem>
-        <primitives.SelectSeparator />
-        {collections.map((c) => (
-          <primitives.SelectItem key={c.id} value={c.name}>
-            {c.content?.title ?? c.name}
-          </primitives.SelectItem>
-        ))}
-      </primitives.Select>
-    </div>
-  );
-}
