--- ee/sites/src/routes/$siteName.submissions._index/ClassicSubmissionsRedirect.tsx
@@ -1,16 +1,0 @@
-import { Link } from 'react-router';
-
-export function ClassicSubmissionsRedirect({ siteName }: { siteName: string }) {
-  return (
-    <div className="mt-4 text-sm text-muted-foreground">
-      Something look wrong?{' '}
-      <Link
-        to={`/app/sites/${siteName}/submissions-classic`}
-        className="text-blue-600 underline hover:text-blue-800"
-      >
-        Return to classic submissions page
-      </Link>
-      .
-    </div>
-  );
-}
