+++ ee/sites/src/routes/$siteName.submissions._index/site-context.format.server.ts
@@ -1,8 +1,8 @@
+import type { SiteContext } from '@curvenote/scms-server';
 import type { SiteLayoutSite } from '../$siteName/layout.format.server.js';
 import { formatSiteLayoutSite } from '../$siteName/layout.format.server.js';
-import type { SiteContext } from '@curvenote/scms-server';
 
-/** Site fields passed into SubmissionList / actions on this route. */
+/** Site fields passed into submission listing UI and actions. */
 export type SubmissionListingSiteContext = SiteLayoutSite;
 
 export function formatSubmissionListingSiteContext(ctx: SiteContext): SubmissionListingSiteContext {
